+++ variables.tf
@@ -76,7 +76,7 @@ variable "lambda_mem_size" {
 
 variable "lambda_runtime" {
   type        = string
-  default     = "python3.9"
+  default     = "python3.12"
   description = "Lambda runtime to use for the function."
 }