+++ versions.tf
@@ -1,7 +1,13 @@
 terraform {
   required_version = "~>  1.0"
   required_providers {
-    aws     = "~>5.0"
-    archive = "~>2.0"
+    aws = {
+      source  = "hashicorp/aws"
+      version = ">= 5.0"
+    }
+    archive = {
+      source  = "hashicorp/archive"
+      version = "~> 2.0"
+    }
   }
 }